+++ src/main/java/com/clova/anifriends/domain/shelter/repository/ShelterRepository.java
@@ -1,7 +1,7 @@
 package com.clova.anifriends.domain.shelter.repository;
 
 import com.clova.anifriends.domain.shelter.Shelter;
-import com.clova.anifriends.domain.shelter.wrapper.ShelterEmail;
+import com.clova.anifriends.domain.shelter.vo.ShelterEmail;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
