+++ src/main/java/com/clova/anifriends/domain/shelter/vo/ShelterPhoneNumberInfo.java
@@ -1,4 +1,4 @@
-package com.clova.anifriends.domain.shelter.wrapper;
+package com.clova.anifriends.domain.shelter.vo;
 
 import com.clova.anifriends.domain.shelter.exception.ShelterBadRequestException;
 import com.clova.anifriends.global.exception.ErrorCode;
